examples/basic/frameworks/haystack_deep_research_agent/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64", "setuptools-scm>=8"]

[tool.setuptools_scm]
root = "../../../.."

[project]
name = "aiq_haystack_deep_research_agent"
dynamic = ["version"]
dependencies = [
  "aiqtoolkit~=1.2",
  "haystack-ai>=2.17.0,<2.19",
  "opensearch-haystack~=4.2.0",
  "nvidia-haystack~=0.3.0",
  "trafilatura~=2.0.0",
  "pypdf~=5.8.0",
  "docstring-parser~=0.16",
]
requires-python = ">=3.11,<3.13"
description = "Haystack Deep Research Agent workflow for AIQ toolkit"
classifiers = ["Programming Language :: Python"]

[tool.uv.sources]
aiqtoolkit = { path = "../../../..", editable = true }

[project.entry-points.'aiq.components']
aiq_haystack_deep_research_agent = "aiq_haystack_deep_research_agent.register"